requirements.txt
streamlit==1.1.0
python-dotenv==0.19.2
azure-storage-blob==12.8.1